Finish adding functinality to all buttons.
there may be some bugs or other things to improve, I have not checked
everything, but this project really doesn't interest me much and I'm
just doing it to complete the odin project path. Im not saying it
doesn't work. it acomplishes everything it is supposed to but yeah,
it isn't my most rigorous project. cheerio